Don't wrap `fnameescape` in a ruby function, causes problems in win32
Incredibly weird behavior: Calling VIM::evaluate("fnameescape"...) directly works
fine but when wrapped in a ruby function it double whacks backslashes. Must
be an issue in vim's ruby calling code, or is it something with ruby itself?
No time to get into that right now, but this works on windows. Needs test on Linux.